Introduce QuicStreamException and QuicStreamResetException.
Motivation:

Currently Quic exception system is flat with a single exception and discards important information that the application protocol (e.g. HTTP/3) cares about.

Changes:

Introduce a QuicStreamException that signals the connection received a STREAM_RESET frame, this provides as well the application protocol error code sent by the remote endpoint.

A QuicStreamException is also introduced that is reused for STREAM_STOPPED frames. Later this might become a subclass as well since this frame also conveys an application protocol error code.

+++ codec-classes-quic/src/main/java/io/netty/handler/codec/quic/QuicheQuicChannel.java
@@ -957,7 +957,7 @@ final class QuicheQuicChannel extends AbstractChannel implements QuicChannel {
         if (recvLen == Quiche.QUICHE_ERR_DONE) {
             return StreamRecvResult.DONE;
         } else if (recvLen < 0) {
-            throw Quiche.convertToException(recvLen, (int) errorCode);
+            throw Quiche.convertToException(recvLen, errorCode);
         }
         buffer.writerIndex(writerIndex + recvLen);
         return finBuffer.getBoolean(0) ? StreamRecvResult.FIN : StreamRecvResult.OK;
